Refactor passes to avoid trait cycles in preparation for async blocks
