+++ compiler/passes/src/const_propagation/program.rs
@@ -17,6 +17,7 @@
 use super::ConstPropagationVisitor;
 
 use leo_ast::{
+    AstReconstructor as _,
     ConstParameter,
     Function,
     Input,
@@ -25,8 +26,6 @@ use leo_ast::{
     ProgramReconstructor,
     ProgramScope,
     Statement,
-    StatementReconstructor as _,
-    TypeReconstructor,
 };
 
 impl ProgramReconstructor for ConstPropagationVisitor<'_> {
