+++ compiler/passes/src/dead_code_elimination/mod.rs
@@ -54,11 +54,9 @@ use crate::Pass;
 use leo_ast::ProgramReconstructor as _;
 use leo_errors::Result;
 
-mod expression;
+mod ast;
 
 mod program;
-
-mod statement;
 
 mod visitor;
 use visitor::*;
