+++ compiler/passes/src/flattening/mod.rs
@@ -57,11 +57,9 @@ use leo_ast::ProgramReconstructor as _;
 use leo_errors::Result;
 use leo_span::Symbol;
 
-mod expression;
+mod ast;
 
 mod program;
-
-mod statement;
 
 mod visitor;
 use visitor::*;
