+++ compiler/passes/src/loop_unrolling/ast.rs
@@ -15,11 +15,29 @@
 // along with the Leo library. If not, see <https://www.gnu.org/licenses/>.
 
 use leo_ast::{Expression::Literal, *};
+
 use leo_errors::LoopUnrollerError;
 
 use super::UnrollingVisitor;
 
-impl StatementReconstructor for UnrollingVisitor<'_> {
+impl AstReconstructor for UnrollingVisitor<'_> {
+    type AdditionalOutput = ();
+
+    /* Expressions */
+    fn reconstruct_repeat(&mut self, input: RepeatExpression) -> (Expression, Self::AdditionalOutput) {
+        // Because the value of `count` affects the type of a repeat expression, we need to assign a new ID to the
+        // reconstructed `RepeatExpression` and update the type table accordingly.
+        let new_id = self.state.node_builder.next_id();
+        let new_count = self.reconstruct_expression(input.count).0;
+        let el_ty = self.state.type_table.get(&input.expr.id()).expect("guaranteed by type checking");
+        self.state.type_table.insert(new_id, Type::Array(ArrayType::new(el_ty, new_count.clone())));
+        (
+            RepeatExpression { expr: self.reconstruct_expression(input.expr).0, count: new_count, id: new_id, ..input }
+                .into(),
+            Default::default(),
+        )
+    }
+
     fn reconstruct_block(&mut self, mut input: Block) -> (Block, Self::AdditionalOutput) {
         self.in_scope(input.id(), |slf| {
             input.statements = input.statements.into_iter().map(|stmt| slf.reconstruct_statement(stmt).0).collect();
