+++ compiler/passes/src/processing_script/ast.rs
@@ -16,12 +16,13 @@
 
 use super::ProcessingScriptVisitor;
 
-use leo_ast::{CallExpression, Expression, ExpressionReconstructor, Location, Variant};
+use leo_ast::{AstReconstructor, CallExpression, Expression, Location, Variant};
 use leo_errors::TypeCheckerError;
 
-impl ExpressionReconstructor for ProcessingScriptVisitor<'_> {
+impl AstReconstructor for ProcessingScriptVisitor<'_> {
     type AdditionalOutput = ();
 
+    /* Expressions */
     fn reconstruct_call(&mut self, input: CallExpression) -> (Expression, Self::AdditionalOutput) {
         if !matches!(self.current_variant, Variant::Script) {
             let callee_program = input.program.unwrap_or(self.program_name);
