--- compiler/passes/src/processing_script/statement.rs
@@ -1,21 +1,0 @@
-// Copyright (C) 2019-2025 Provable Inc.
-// This file is part of the Leo library.
-
-// The Leo library is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// The Leo library is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with the Leo library. If not, see <https://www.gnu.org/licenses/>.
-
-use super::ProcessingScriptVisitor;
-
-use leo_ast::StatementReconstructor;
-
-impl StatementReconstructor for ProcessingScriptVisitor<'_> {}
