+++ compiler/passes/src/write_transforming/ast.rs
@@ -15,29 +15,7 @@
 // along with the Leo library. If not, see <https://www.gnu.org/licenses/>.
 
 use super::WriteTransformingVisitor;
-
-use leo_ast::{
-    ArrayAccess,
-    ArrayExpression,
-    AssignStatement,
-    AssociatedFunctionExpression,
-    BinaryExpression,
-    CallExpression,
-    CastExpression,
-    Expression,
-    ExpressionReconstructor,
-    Identifier,
-    MemberAccess,
-    Node,
-    Statement,
-    StructExpression,
-    StructVariableInitializer,
-    TernaryExpression,
-    TupleAccess,
-    TupleExpression,
-    Type,
-    UnaryExpression,
-};
+use leo_ast::*;
 use leo_span::Symbol;
 
 impl WriteTransformingVisitor<'_> {
@@ -59,9 +37,10 @@ impl WriteTransformingVisitor<'_> {
     }
 }
 
-impl ExpressionReconstructor for WriteTransformingVisitor<'_> {
+impl AstReconstructor for WriteTransformingVisitor<'_> {
     type AdditionalOutput = Vec<Statement>;
 
+    /* Expressions */
     fn reconstruct_identifier(&mut self, input: Identifier) -> (Expression, Self::AdditionalOutput) {
         let ty = self.state.type_table.get(&input.id()).unwrap();
         let mut statements = Vec::new();
@@ -258,5 +237,104 @@ impl ExpressionReconstructor for WriteTransformingVisitor<'_> {
 
     fn reconstruct_unit(&mut self, input: leo_ast::UnitExpression) -> (Expression, Self::AdditionalOutput) {
         (input.into(), Default::default())
+    }
+
+    /* Statements */
+    /// This is the only reconstructing function where we do anything other than traverse and combine statements,
+    /// by calling `reconstruct_assign_place` and `reconstruct_assign_recurse`.
+    fn reconstruct_assign(&mut self, input: AssignStatement) -> (Statement, Self::AdditionalOutput) {
+        let (value, mut statements) = self.reconstruct_expression(input.value);
+        let place = self.reconstruct_assign_place(input.place);
+        self.reconstruct_assign_recurse(place, value, &mut statements);
+        (Statement::dummy(), statements)
+    }
+
+    fn reconstruct_assert(&mut self, input: leo_ast::AssertStatement) -> (Statement, Self::AdditionalOutput) {
+        let mut statements = Vec::new();
+        let stmt = AssertStatement {
+            variant: match input.variant {
+                AssertVariant::Assert(expr) => {
+                    let (expr, statements2) = self.reconstruct_expression(expr);
+                    statements.extend(statements2);
+                    AssertVariant::Assert(expr)
+                }
+                AssertVariant::AssertEq(left, right) => {
+                    let (left, statements2) = self.reconstruct_expression(left);
+                    statements.extend(statements2);
+                    let (right, statements3) = self.reconstruct_expression(right);
+                    statements.extend(statements3);
+                    AssertVariant::AssertEq(left, right)
+                }
+                AssertVariant::AssertNeq(left, right) => {
+                    let (left, statements2) = self.reconstruct_expression(left);
+                    statements.extend(statements2);
+                    let (right, statements3) = self.reconstruct_expression(right);
+                    statements.extend(statements3);
+                    AssertVariant::AssertNeq(left, right)
+                }
+            },
+            ..input
+        }
+        .into();
+        (stmt, Default::default())
+    }
+
+    fn reconstruct_block(&mut self, block: Block) -> (Block, Self::AdditionalOutput) {
+        let mut statements = Vec::with_capacity(block.statements.len());
+
+        // Reconstruct the statements in the block, accumulating any additional statements.
+        for statement in block.statements {
+            let (reconstructed_statement, additional_statements) = self.reconstruct_statement(statement);
+            statements.extend(additional_statements);
+            if !reconstructed_statement.is_empty() {
+                statements.push(reconstructed_statement);
+            }
+        }
+
+        (Block { statements, ..block }, Default::default())
+    }
+
+    fn reconstruct_definition(&mut self, mut input: DefinitionStatement) -> (Statement, Self::AdditionalOutput) {
+        let (value, mut statements) = self.reconstruct_expression(input.value);
+        input.value = value;
+        match input.place.clone() {
+            DefinitionPlace::Single(identifier) => {
+                statements.push(input.into());
+                self.define_variable_members(identifier, &mut statements);
+            }
+            DefinitionPlace::Multiple(identifiers) => {
+                statements.push(input.into());
+                for &identifier in identifiers.iter() {
+                    self.define_variable_members(identifier, &mut statements);
+                }
+            }
+        }
+        (Statement::dummy(), statements)
+    }
+
+    fn reconstruct_expression_statement(&mut self, input: ExpressionStatement) -> (Statement, Self::AdditionalOutput) {
+        let (expression, statements) = self.reconstruct_expression(input.expression);
+        (ExpressionStatement { expression, ..input }.into(), statements)
+    }
+
+    fn reconstruct_iteration(&mut self, _input: IterationStatement) -> (Statement, Self::AdditionalOutput) {
+        panic!("`IterationStatement`s should not be in the AST at this point.");
+    }
+
+    fn reconstruct_return(&mut self, input: ReturnStatement) -> (Statement, Self::AdditionalOutput) {
+        let (expression, statements) = self.reconstruct_expression(input.expression);
+        (ReturnStatement { expression, ..input }.into(), statements)
+    }
+
+    fn reconstruct_conditional(&mut self, input: leo_ast::ConditionalStatement) -> (Statement, Self::AdditionalOutput) {
+        let (condition, mut statements) = self.reconstruct_expression(input.condition);
+        let (then, statements2) = self.reconstruct_block(input.then);
+        statements.extend(statements2);
+        let otherwise = input.otherwise.map(|oth| {
+            let (expr, statements3) = self.reconstruct_statement(*oth);
+            statements.extend(statements3);
+            Box::new(expr)
+        });
+        (ConditionalStatement { condition, then, otherwise, ..input }.into(), statements)
     }
 }
