+++ compiler/passes/src/write_transforming/mod.rs
@@ -20,11 +20,9 @@ use leo_ast::ProgramReconstructor as _;
 
 use leo_errors::Result;
 
-mod expression;
+mod ast;
 
 mod program;
-
-mod statement;
 
 mod visitor;
 use visitor::*;
